backend/requirements.txt
### requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6
Pillow==10.0.0
cryptography==41.0.4
pyotp==2.9.0
qrcode==7.4.2
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
marshmallow==3.20.1
bleach==6.0.0
email-validator==2.0.0